Fixes duplicate lines when a curve has a straight section
A line with a straight section, then a curve that contains a section that could be interpreted as a straight segment, will get interpreted as a straight section, another straight section, and then a curve starting from the start of the second straight section. This was the case for CS6 at least, and was making the meshification process going on in RageTools SVG-In for Unity3D fail.

This removes all segments from a line where the line is not a continuous line. It's a bit of a brute force approach, but it was giving me more correct results than before the change (and I don't know how to detect this case properly)